+++ src/main/java/com/example/ililbooks/client/book/dto/BookApiWrapper.java
@@ -1,0 +1,7 @@
+package com.example.ililbooks.client.book.dto;
+
+public record BookApiWrapper (
+        BookApiResponse [] result
+){
+
+}
