--- src/main/java/com/example/ililbooks/client/dto/BookApiWrapper.java
@@ -1,9 +1,0 @@
-package com.example.ililbooks.client.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record BookApiWrapper (
-        BookApiResponse [] result
-){
-
-}
