+++ src/main/java/com/example/ililbooks/client/naver/dto/NaverApiProfileWrapper.java
@@ -1,4 +1,4 @@
-package com.example.ililbooks.client.dto;
+package com.example.ililbooks.client.naver.dto;
 
 public record NaverApiProfileWrapper(
         NaverApiProfileResponse response
